Fix CDI initialization of root-span fields and fix root-span for golden signals
GoldenSignalsRootSpanFields was iterating all metrics, not just the ones from the TrafficClassifier. This change will rely on GoldenSignalsFilter to classify the traffic and store the function list in ThreadContext, then pull them out and iterate them here to construct the traffic_type field.

DefaultHoneycombManager was trying to init the rootSpan fields provider list during ctor execution, which didn't play well with CDI in the Indy monolith. This change moves that logic to the init() method, which runs @PostConstruct.