Refining honeycomb and otel tracers, adding jhttpc wrapper
The jHTTPc helper is for tracing, and wraps the HttpFactory, HttpClient, and HttpResponse in order to
manage the client-side span each time a request is made. This will record the duration automatically
but also records the target URL and the response status.

The changes to the Honeycomb tracer are mainly to restore thread span nesting, but also implement
more context propagation convenience methods, and improve propagation into and out of thread spans.

Likewise, the otel tracer changes mainly handle context propagation and propagation of fields into and
out of thread spans.