Cache answers for Prometheus expressed metrics + Honeycomb sampling
Also, reduce log levels in some places to TRACE, and add a metric to the golden signals that is just
a histogram of raw latency nanoseconds, to help find very small time differences that can add up to huge
performance penalties.

Adding a Histogram type to O11yphant's API to help with reporting things like that.