Add ability to delay span close
We sometimes need to delay span close until all registered close-blockers call the close() method. This is important in cases like Indy's download function, where the servlet/ REST API may return a response to the client that includes a TransferStreamingOutput. This output class will continue to interact with the client after the REST call runs its finally block. In current situations, this means call durations don't account for all of the transfer time.

This change will register a series of close-blocking actors, associated with a set of field injectors. When the close() call is executed the same number of times as the registered close-blockers, all fields are added and the span is actually closed.

**NOTE:** This may not work properly with ThreadLocal registration of active spans...we need something that keeps a thread context separate from any given thread, so it's safe to use with durable thread pools.