Improve pausing of latency measurements and fix threaded traces
Threaded traces needed a queue of SpanAdapters in TraceManager, since threads sometimes create child spans. Additionally, thread context transfer back to the calling thread must include updating all 'add.*' and 'cumulative*' in-progress fields.

In that vein, this also adds a notion of 'add.latency_pause' which allows us to accumulate pause time while transfers from upstream or to/from the client finish. These times are dependent on client and network conditions, so should not be factored into Indy's own performance...though it's also possible Indy's disk speed may factor into that pause time, and be lost in the resulting latency calculation.

Finally, I'm updating all in-progress fields in spans to be of type Double. This assumes we won't be keeping progress on non-scalar fields, and it uses the widest resolution for scalar fields, so it should be pretty safe.